feat: DataManager 타입 구현
